requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
filelock==3.16.0
Flask==3.0.3
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==1.44.0
openai-whisper==20231117
opencv-python==4.10.0.84
packaging==24.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.0
sympy==1.13.2
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
Pillow==10.3.0
torchvision==0.18.0
transformers==4.40.2

